machine_actuator: Split into separate files
